feat: Refactor playground vue
