+++ .eslintignore
@@ -11,6 +11,7 @@ node_modules
 .eslintcache
 .cache-loader
 .jest-cache
+*.styl
 
 # testing
 coverage
