+++ .prettierignore
@@ -9,6 +9,7 @@ node_modules
 .eslintcache
 .cache-loader
 .jest-cache
+*.styl
 
 # testing
 coverage
