+++ playgrounds/vue/src/inputs/InputCheckboxList.vue
@@ -1,5 +1,5 @@
 <template>
-  <div>
+  <NonLabelWrapper :field="vField">
     <h3>{{ label }}</h3>
     <label v-for="(option, key) in options" :key="key">
       <input
@@ -10,22 +10,19 @@
         :checked="vField.value[key]"
       />{{ option }}</label
     >
-    <p style="min-height: 1.15em">
-      <span v-show="vField.touched && vField.error" style="color: red">{{
-        vField.error
-      }}</span>
-      <span v-show="!vField.error && vField.warning" style="color: orange">{{
-        vField.warning
-      }}</span>
-    </p>
-  </div>
+  </NonLabelWrapper>
 </template>
 
 <script>
 import VueTypes from 'vue-types';
 import FieldMixin from '@validate-me/vue/FieldMixin';
 
+import NonLabelWrapper from './wrappers/NonLabelWrapper';
+
 export default {
+  components: {
+    NonLabelWrapper,
+  },
   mixins: [FieldMixin],
   props: {
     label: VueTypes.string.isRequired,
