+++ playgrounds/vue/src/inputs/InputRadioList.vue
@@ -1,6 +1,5 @@
 <template>
-  <div>
-    <h3>{{ label }}</h3>
+  <NonLabelWrapper :field="vField">
     <label v-for="(option, key) in options" :key="key">
       <input
         v-validate-me
@@ -11,22 +10,19 @@
       />
       {{ option }}
     </label>
-    <p style="min-height: 1.15em">
-      <span v-show="vField.touched && vField.error" style="color: red">{{
-        vField.error
-      }}</span>
-      <span v-show="!vField.error && vField.warning" style="color: orange">{{
-        vField.warning
-      }}</span>
-    </p>
-  </div>
+  </NonLabelWrapper>
 </template>
 
 <script>
 import VueTypes from 'vue-types';
 import FieldMixin from '@validate-me/vue/FieldMixin';
 
+import NonLabelWrapper from './wrappers/NonLabelWrapper';
+
 export default {
+  components: {
+    NonLabelWrapper,
+  },
   mixins: [FieldMixin],
   props: {
     label: VueTypes.string.isRequired,
